Tweaked error reporting; fixed indentation bug.
The call stack wouldn't show the current line in the "main" unless there
was also a function on the call stack. Also, indentation was ignored on
the first line since I only parse indentation when there are newlines,
so I fixed it by adding a newline to the beginning of the file and
offsetting the line numbering by -1 to compensate.